+++ source/extensions/omni.isaac.ros_bridge/docs/CHANGELOG.md
@@ -11,6 +11,7 @@ Fixed
 - horizontal and vertical aperture use camera prim values instead of computing vertical aperture
 - lidar components publish point cloud data as PCL2 messages instead of PCL
 - lidar PCL2 messages only contain points that hit
+- lidar publisher publishes a full scan for point cloud data
 
 Added
 -------